JodoroidApplicationからAPI関連のメソッドを切り出す
